Stop using express
We currently use express to handle two URLs - `POST /init` and
`POST /run`, and use the middleware functionality to parse JSON bodies.

This is not necessary, as this kind of simple routing can be handled
with an `if` statement.

This is done in interest of reducing code size and both initialization
and run overhead.